testing: seperating out into a library for now. Need to add architecture subdirectories among alot of other fixes.
