+++ examples/microkit_sddf_serial/meta.py
@@ -49,7 +49,7 @@ def generate(sdf_file: str, output_dir: str, dtb: DeviceTree):
 
     mr = MemoryRegion("mr1", 0x1000)
     sdf.add_mr(mr)
-    map = Map(mr, 0x600000, "rw")
+    map = Map(mr, 0x900000, "rw")
     debugger.add_map(map)
 
     ping = ProtectionDomain("ping", "ping.elf", priority=96)
